Implement connection attributes
- Add support for connection attributes
- Clarify typing and purpose of attributes
- Support reporting attributes where only a single value are permitted
- Report when attributes are not supported

+++ flight_sql/flight_sql_auth_method.cc
@@ -61,7 +61,7 @@ public:
     if (login_timeout.has_value()) {
       // ODBC's LOGIN_TIMEOUT attribute and FlightCallOptions.timeout use
       // seconds as time unit.
-      double timeout_seconds = boost::get<double>(login_timeout.value());
+      double timeout_seconds = static_cast<double>(boost::get<uint32_t>(login_timeout.value()));
       if (timeout_seconds > 0) {
         auth_call_options.timeout = TimeoutDuration{timeout_seconds};
       }
